Added first Get methods for Articles, ArticlesDao and ArticlesServices

+++ src/main/java/it/rdev/blog/api/dao/TagDao.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import it.rdev.blog.api.dao.entity.Tag;
 
 @Repository
-public interface TagDao extends CrudRepository<Tag,Integer> {
+public interface TagDao extends CrudRepository<Tag,Long> {
 	
 	@Query("Select t From Tag t")
 	Set<Tag> findAll();
